kexec: Introduce default_initrd_fd to pass internal initrd information
At present, initrd information is passed through --initrd args, but for
the UKI case, the initrd may be stored in the .initrd section.

It means that initrd is perceived at the probe stage. And there should
be a way to carry this information forward to the load stage. In oder to
implement that, introducing a global default_initrd_fd to do it.

Signed-off-by: Pingfan Liu <piliu@redhat.com>
Signed-off-by: Simon Horman <horms@kernel.org>